Confirm leave
* Fix rebase issues

* Fix rebase issues

* Fix rebase issues

* Fix rebase issues

* Fix merge conflicts

* Fix merge conflicts

* Fix merge conflicts

* Fix merge conflicts

* Fix main.js merge conflicts
fj;
;

* fix main.js merge conflicts

* Fix main.js

* Fix main.js

* message routing routed

* confirm to leave group

+++ client/src/containers/leave_group.js
@@ -13,9 +13,11 @@ class LeaveGroup extends Component {
   }
 
   onHandleLeaveGroup() {
-    this.props.leaveGroup(this.props.group.id, this.props.group.profile_id);
-    // this.props.handleDeleteGroup();
-    this.props.handleDeleteGroup();
+    var confirmDelete = confirm('Are you sure you want to leave the group?');
+    if (confirmDelete) {
+      this.props.leaveGroup(this.props.group.id, this.props.group.profile_id);
+      this.props.handleDeleteGroup();
+    }
   }
 
   render() {
